+++ sample-android-lib/src/main/java/me/tatarka/retrolambda/sample/lib/Lib.java
@@ -4,9 +4,6 @@ import android.support.annotation.IntDef;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-/**
- * Created by evan on 3/29/15.
- */
 public class Lib {
 
     @Retention(RetentionPolicy.SOURCE)