+++ sample-java8-lib/src/main/java/me/tatarka/retrolambda/sample/javalib/Function.java
@@ -1,0 +1,5 @@
+package me.tatarka.retrolambda.sample.javalib;
+
+public interface Function {
+    String run();
+}