+++ sample-java8-lib/src/main/java/me/tatarka/retrolambda/sample/javalib/JavaLib.java
@@ -1,0 +1,12 @@
+package me.tatarka.retrolambda.sample.javalib;
+
+public class JavaLib {
+    public static Function getHello() {
+        return new Function() {
+            @Override
+            public String run() {
+                return "Hello, Retrolambda (from java lib)!";
+            }
+        };
+    }
+}