+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#### 3.7.0-SNAPSHOT
+- Added support for processing libraries with `retrolambda.processLibraries true` on android projects. This
+requires retrolambda 2.6.0-SNAPSHOT. It will work with external jars/aars as well as java/android modules.
+
 #### 3.6.0
 - *Deprecated* oldJdk (and reading from JAVA6/JAVA7_HOME).
 - *Deprecated* running gradle with java 6 or 7.